update kairosdb supervisord conf to ensure process runs for at least 10secs
Kairosdb takes a while to start up.  If Cassandra is not yet ready,
which is common when launching with docker in another container, then
the kairosdb proces can fail to start.  This commit instructs
supervisord to wait 10secs before conforming that the process
started successfully so that it can continue to restart until
cassandra is available.