allow symlinking <code dir>/$project to a directory on your system
so you can for example have raintank_code/grafana be a symlink
to the directory containing the code on your system.
develop what's on the dev container using your familiar dev environment!

to make this happen, we need to volume map the individual project
directories, because otherwise the container would see the symlink to
a non-existent (in the container) directory.

also you should use a .notouch in your code dir so this script
doesn't automatically try to fetch and clobber your code dir.