fix grafana plugins support. fix #79
1) support github url basename != dirname
2) ... which we then use for grafana plugins in raintank_code

Note: I opted to keep every directory within raintank_code to be 1
code directory: I didn't want to have a hiearchy (e.g. a plugins dir
with code dirs within it), because that would be non-obvious for
people glancing in the raintank_code dir.
That means there isn't a `plugins` dir anymore to volume-bind with
docker, so we need to map plugins individually in the docker config.
This means that we could technically forgo 1) and 2) and just have
a directory raintank_code/worldping-app just like the others,

however:
* over time we may want to add additional Grafana plugins, which are
  only guaranteed to be unique using their grafana slug, including org
  name
* it seems cleaner to abstract away the assurecode function anyway